Move `__global_logging_exception_handler` to `salt._logging`
